requirements.txt
appdirs==1.4.4 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
autopage==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
cliff==4.3.0 ; python_version >= "3.10" and python_version < "4.0"
cmd2==2.4.3 ; python_version >= "3.10" and python_version < "4.0"
cryptography==41.0.1 ; python_version >= "3.10" and python_version < "4.0"
debtcollector==2.5.0 ; python_version >= "3.10" and python_version < "4.0"
decorator==5.1.1 ; python_version >= "3.10" and python_version < "4.0"
dogpile-cache==1.2.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
importlib-metadata==6.6.0 ; python_version >= "3.10" and python_version < "4.0"
iso8601==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
jsonpatch==1.32 ; python_version >= "3.10" and python_version < "4.0"
jsonpointer==2.3 ; python_version >= "3.10" and python_version < "4.0"
keystoneauth1==5.2.0 ; python_version >= "3.10" and python_version < "4.0"
msgpack==1.0.5 ; python_version >= "3.10" and python_version < "4.0"
netaddr==0.8.0 ; python_version >= "3.10" and python_version < "4.0"
netifaces==0.11.0 ; python_version >= "3.10" and python_version < "4.0"
openstacksdk==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
os-service-types==1.7.0 ; python_version >= "3.10" and python_version < "4.0"
osc-lib==2.8.0 ; python_version >= "3.10" and python_version < "4.0"
oslo-config==9.1.1 ; python_version >= "3.10" and python_version < "4.0"
oslo-i18n==6.0.0 ; python_version >= "3.10" and python_version < "4.0"
oslo-serialization==5.1.1 ; python_version >= "3.10" and python_version < "4.0"
oslo-utils==6.1.0 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.10" and python_version < "4.0"
pbr==5.11.1 ; python_version >= "3.10" and python_version < "4.0"
prettytable==3.8.0 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.10" and python_version < "4.0"
pyperclip==1.8.2 ; python_version >= "3.10" and python_version < "4.0"
pyreadline3==3.4.1 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
python-cinderclient==9.3.0 ; python_version >= "3.10" and python_version < "4.0"
python-keystoneclient==5.1.0 ; python_version >= "3.10" and python_version < "4.0"
python-novaclient==18.3.0 ; python_version >= "3.10" and python_version < "4.0"
python-openstackclient==5.5.0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
requestsexceptions==1.4.0 ; python_version >= "3.10" and python_version < "4.0"
rfc3986==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
simplejson==3.19.1 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
stevedore==5.1.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.0.3 ; python_version >= "3.10" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.10" and python_version < "4.0"
wrapt==1.15.0 ; python_version >= "3.10" and python_version < "4.0"
zipp==3.15.0 ; python_version >= "3.10" and python_version < "4.0"
